Fix all pending issues
