+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   font_awesome_flutter: ^8.8.1
   shared_preferences: ^0.5.6+3
   scrollable_positioned_list: ^0.1.4
+  scroll_app_bar: ^0.2.0+1
 
 dev_dependencies:
   flutter_test:
